All code completed. Missing rotation functionality. Has not been tested on tablet devices

+++ app/src/main/java/com/krkp/menumaker/database/entities/Orders.kt
@@ -12,5 +12,5 @@ data class Orders (
     val dateTime: String,
     val username: String,
     val order: String,
-    val total: Int
+    val total: Double
 ): Parcelable
